Ported to 1.16.5
