+++ .vscode/settings.json
@@ -1,5 +1,4 @@
 {
-    "git.ignoreLimitWarning": true,
-    "java.dependency.packagePresentation": "hierarchical",
-    "java.configuration.updateBuildConfiguration": "automatic"
+    "java.configuration.updateBuildConfiguration": "automatic",
+    "java.dependency.packagePresentation": "hierarchical"
 }
