--- src/main/java/dev/blythefn/sorgscraft/item/custom/Curry.java
@@ -1,52 +1,0 @@
-package dev.blythefn.sorgscraft.item.custom;
-
-import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResultHolder;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.ItemUtils;
-import net.minecraft.world.item.UseAnim;
-import net.minecraft.world.level.Level;
-
-public class Curry extends Item {
-
-    public Curry(Properties pProperties) {
-        super(pProperties);
-    }
-    
-    public ItemStack finishUsingItem(ItemStack pStack, Level pLevel, LivingEntity pEntityLiving) {
-        if (!pLevel.isClientSide) {
-            if(pEntityLiving instanceof Player)
-                {
-                    ((Player)pEntityLiving).setSecondsOnFire(20);
-                }
-        };
-
-        if (pEntityLiving instanceof Player && !((Player)pEntityLiving).getAbilities().instabuild) {
-        pStack.shrink(1);
-        }
-
-        return pStack;
-    }
-
-    public int getUseDuration(ItemStack pStack) {
-        return 32;
-    }
-
-    public UseAnim getUseAnimation(ItemStack pStack) {
-        return UseAnim.EAT;
-    }
-
-    public InteractionResultHolder<ItemStack> use(Level pLevel, Player pPlayer, InteractionHand pHand) {
-        return ItemUtils.startUsingInstantly(pLevel, pPlayer, pHand);
-    }
-
-    @Override
-    public boolean hurtEnemy(ItemStack pStack, LivingEntity pTarget, LivingEntity pAttacker) {
-        pTarget.setSecondsOnFire(20);
-        return super.hurtEnemy(pStack, pTarget, pAttacker);
-    }
-
-}
