--- src/main/java/dev/blythefn/sorgscraft/item/custom/Vape.java
@@ -1,45 +1,0 @@
-package dev.blythefn.sorgscraft.item.custom;
-
-import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResultHolder;
-import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraft.world.effect.MobEffects;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.ItemUtils;
-import net.minecraft.world.item.UseAnim;
-import net.minecraft.world.level.Level;
-
-public class Vape extends Item {
-
-    public Vape(Properties pProperties) {
-        super(pProperties);
-    }
-
-    public ItemStack finishUsingItem(ItemStack pStack, Level pLevel, LivingEntity pEntityLiving) {
-        if (!pLevel.isClientSide) {
-            pEntityLiving.addEffect(new MobEffectInstance(MobEffects.CONFUSION, 600, 0, true, false, false));
-            pEntityLiving.addEffect(new MobEffectInstance(MobEffects.LEVITATION, 600, 0, true, false, false));
-            pStack.hurtAndBreak(1, ((Player)pEntityLiving), p -> {
-                p.broadcastBreakEvent(pEntityLiving.getUsedItemHand());
-            });
-        };
-
-        return pStack;
-    }
-
-    public int getUseDuration(ItemStack pStack) {
-        return 32;
-    }
-
-    public UseAnim getUseAnimation(ItemStack pStack) {
-        return UseAnim.DRINK;
-    }
-
-    public InteractionResultHolder<ItemStack> use(Level pLevel, Player pPlayer, InteractionHand pHand) {
-        return ItemUtils.startUsingInstantly(pLevel, pPlayer, pHand);
-    }
-    
-}
