+++ src/main/java/dev/unchilled/sorgscraft/item/ModArmorMaterials.java
@@ -1,17 +1,15 @@
-package dev.blythefn.sorgscraft.item;
+package dev.unchilled.sorgscraft.item;
 import java.util.function.Supplier;
 
-import dev.blythefn.sorgscraft.sorgscraft;
-import net.minecraft.util.LazyLoadedValue;
-import net.minecraft.sounds.SoundEvent;
-import net.minecraft.sounds.SoundEvents;
-import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.item.ArmorMaterial;
-import net.minecraft.world.item.crafting.Ingredient;
+import dev.unchilled.sorgscraft.sorgscraft;
+import net.minecraft.util.LazyValue;
+import net.minecraft.util.SoundEvent;
+import net.minecraft.util.SoundEvents;
+import net.minecraft.inventory.EquipmentSlotType;
+import net.minecraft.item.IArmorMaterial;
+import net.minecraft.item.crafting.Ingredient;
 
-@SuppressWarnings("deprecation")
-
-public enum ModArmorMaterials implements ArmorMaterial {
+public enum ModArmorMaterials implements IArmorMaterial {
     GAP("gap", 37, new int[]{5, 7, 9, 3}, 12, SoundEvents.ARMOR_EQUIP_IRON, 1.0f, 0.0f, () -> {
         return Ingredient.of(Moditems.TOOTH_OF_THE_GOBLIN.get());
     });
@@ -24,7 +22,7 @@ public enum ModArmorMaterials implements ArmorMaterial {
     private final SoundEvent sound;
     private final float toughness;
     private final float knockbackResistance;
-    private final LazyLoadedValue<Ingredient> repairIngredient;
+    private final LazyValue<Ingredient> repairIngredient;
  
     private ModArmorMaterials(String p_40474_, int p_40475_, int[] p_40476_, int p_40477_, SoundEvent p_40478_, float p_40479_, float p_40480_, Supplier<Ingredient> p_40481_) {
        this.name = p_40474_;
@@ -34,14 +32,14 @@ public enum ModArmorMaterials implements ArmorMaterial {
        this.sound = p_40478_;
        this.toughness = p_40479_;
        this.knockbackResistance = p_40480_;
-       this.repairIngredient = new LazyLoadedValue<>(p_40481_);
+       this.repairIngredient = new LazyValue<>(p_40481_);
     }
 
-    public int getDurabilityForSlot(EquipmentSlot pSlot) {
+    public int getDurabilityForSlot(EquipmentSlotType pSlot) {
         return HEALTH_PER_SLOT[pSlot.getIndex()] * this.durabilityMultiplier;
     }
 
-    public int getDefenseForSlot(EquipmentSlot pSlot) {
+    public int getDefenseForSlot(EquipmentSlotType pSlot) {
         return this.slotProtections[pSlot.getIndex()];
     }
