+++ src/main/java/dev/unchilled/sorgscraft/item/ModCreativeModeTab.java
@@ -1,0 +1,15 @@
+package dev.unchilled.sorgscraft.item;
+
+import net.minecraft.item.ItemGroup;
+import net.minecraft.item.ItemStack;
+
+public class ModCreativeModeTab {
+    public static final ItemGroup SORGS_TAB = new ItemGroup("SorgsCraftTab") {
+        @Override
+        public ItemStack makeIcon() {
+            return new ItemStack(Moditems.THE_GAP.get());
+        }
+    };
+}
+
+
