+++ src/main/java/dev/unchilled/sorgscraft/item/ModItemTier.java
@@ -1,14 +1,12 @@
-package dev.blythefn.sorgscraft.item;
+package dev.unchilled.sorgscraft.item;
 
 import java.util.function.Supplier;
 
-import net.minecraft.util.LazyLoadedValue;
-import net.minecraft.world.item.Tier;
-import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.util.LazyValue;
+import net.minecraft.item.IItemTier;
+import net.minecraft.item.crafting.Ingredient;
 
-@SuppressWarnings( "deprecation" )
-
-public enum ModItemTier implements Tier {
+public enum ModItemTier implements IItemTier {
 
     ROB(3, 500, 8.0F, -5.0F, 10, () -> {
         return Ingredient.of(Moditems.TINDER.get());
@@ -19,7 +17,7 @@ public enum ModItemTier implements Tier {
     private final float speed;
     private final float damage;
     private final int enchantmentValue;
-    private final LazyLoadedValue<Ingredient> repairIngredient;
+    private final LazyValue<Ingredient> repairIngredient;
 
     ModItemTier(int harvestLevel, int maxUses, float efficiency, float attackDamage, int enchantmentValue, Supplier<Ingredient> repairIngredient) {
         this.level = harvestLevel;
@@ -27,7 +25,7 @@ public enum ModItemTier implements Tier {
         this.speed = efficiency;
         this.damage = attackDamage;
         this.enchantmentValue = enchantmentValue;
-        this.repairIngredient = new LazyLoadedValue<>(repairIngredient);
+        this.repairIngredient = new LazyValue<>(repairIngredient);
     }
 
     @Override
