+++ src/main/resources/META-INF/mods.toml
@@ -1,58 +1,28 @@
-# This is an example mods.toml file. It contains the data relating to the loading mods.
-# There are several mandatory fields (#mandatory), and many more that are optional (#optional).
-# The overall format is standard TOML format, v0.5.0.
-# Note that there are a couple of TOML lists in this file.
-# Find more information on toml format here:  https://github.com/toml-lang/toml
-# The name of the mod loader type to load - for regular FML @Mod mods it should be javafml
 modLoader="javafml" #mandatory
-# A version range to match for said mod loader - for regular FML @Mod it will be the forge version
-loaderVersion="[37,)" #mandatory This is typically bumped every Minecraft version by Forge. See our download page for lists of versions.
-# The license for you mod. This is mandatory metadata and allows for easier comprehension of your redistributive properties.
-# Review your options at https://choosealicense.com/. All rights reserved is the default copyright stance, and is thus the default here.
-license="All rights reserved"
-# A URL to refer people to when problems occur with this mod
-#issueTrackerURL="https://sorgscraft.unchilled.dev/" #optional
-# A list of mods - how many allowed here is determined by the individual mod loader
-[[mods]] #mandatory
-# The modid of the mod
-modId="sorgscraft" #mandatory
-# The version number of the mod - there's a few well known ${} variables useable here or just hardcode it
-# ${file.jarVersion} will substitute the value of the Implementation-Version as read from the mod's JAR file metadata
-# see the associated build.gradle script for how to populate this completely automatically during a build
-version="1.17.1-0.0.3" #mandatory
- # A display name for the mod
-displayName="SorgsCraft" #mandatory
-# A URL to query for updates for this mod. See the JSON update specification https://mcforge.readthedocs.io/en/latest/gettingstarted/autoupdate/
-#updateJSONURL="https://sorgscraft.unchilled.dev/updates.json/" #optional
-# A URL for the "homepage" for this mod, displayed in the mod UI
-#displayURL="https://sorgscraft.unchilled.dev/" #optional
-# A file name (in the root of the mod JAR) containing a logo for display
-#logoFile="examplemod.png" #optional
-# A text field displayed in the mod UI
-credits="Brane Sorgs, Japan, Egotopia, Blossoms, and all the random ass servers in between" #optional
-# A text field displayed in the mod UI
-authors="Blythe, unchilled" #optional
-# The description text for the mod (multi line!) (#mandatory)
+loaderVersion="[36,)"
+license="GNU GPLv3"
+issueTrackerURL="https://sorgscraft.unchilled.dev/"
+[[mods]]
+modId="sorgscraft"
+version="1.16.5-0.0.4"
+displayName="SorgsCraft"
+updateJSONURL="https://sorgscraft.unchilled.dev/updates.json" 
+displayURL="https://sorgscraft.unchilled.dev" 
+#logoFile="examplemod.png"
+credits="Brane Sorgs, Japan, Egotopia, Blossoms, and all the random ass servers in between"
+authors="Blythe, unchilled"
 description='''
 Adds members from the Discord along with special abilities, items, and blocks from the Discord.
 '''
-# A dependency - use the . to indicate dependency for a specific modid. Dependencies are optional.
 [[dependencies.sorgscraft]] #optional
-    # the modid of the dependency
-    modId="forge" #mandatory
-    # Does this dependency have to exist - if not, ordering below must be specified
-    mandatory=true #mandatory
-    # The version range of the dependency
-    versionRange="[37,)" #mandatory
-    # An ordering relationship for the dependency - BEFORE or AFTER required if the relationship is not mandatory
+    modId="forge"
+    mandatory=true
+    versionRange="[36,)"
     ordering="NONE"
-    # Side this dependency is applied on - BOTH, CLIENT or SERVER
     side="BOTH"
-# Here's another dependency
 [[dependencies.sorgscraft]]
     modId="minecraft"
     mandatory=true
-# This version range declares a minimum of the current minecraft version up to but not including the next major version
-    versionRange="[1.17.1,1.18)"
+    versionRange="[1.16.5,1.17)"
     ordering="NONE"
     side="BOTH"
